added keywordExtract.js for keyword extraction functions
